+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "open-editor-search" extension will be documented in this file.
 
+## [0.0.4] - 2022-03-21
+
+- Open Editor based upon the URI.
+- Fix bug when a markdown preview was open.
+
 ## [0.0.3] - 2022-03-17
 
 - Update label extraction regex to handle unix path - [Pull Request](https://github.com/RobertOstermann/vscode-open-editor-search/pull/1)